Backend/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.0.2
matplotlib==3.5.1
pandas==1.3.4
matplotlib==3.5.1
flask_cors==3.0.10
requests==2.26.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.8.1+cpu           
tensorflow-cpu==2.5.0
transformers==4.14.1
pymongo[srv]